Ignore stop() calls in child process

+++ src/test/mysqld.py
@@ -79,8 +79,10 @@ class Mysqld(object):
         return self
 
     def __exit__(self, *args):
-        self.stop()
-        self.cleanup()
+        import os
+        if self.pid and self._owner_pid == os.getpid():
+            self.stop()
+            self.cleanup()
 
     def cleanup(self):
         from shutil import rmtree
@@ -145,10 +147,13 @@ class Mysqld(object):
             conn.close()
 
     def stop(self, _signal=signal.SIGTERM):
+        import os
         if self.pid is None:
             return  # not started
 
-        import os
+        if self._owner_pid != os.getpid():
+            return  # could not stop in child process
+
         try:
             os.kill(self.pid, _signal)
             while (os.waitpid(self.pid, 0)):
